# api/docs/swagger.yaml
definitions:
  item.AcceptSwapRequestRequest:
    properties:
      swap_id:
        type: string
    type: object
...
  item.AddRatingRequest:
    properties:
      comment:
        type: string
      rater_id:
        type: string
      rating:
        type: number
      swap_id:
        type: string
      user_id:
        type: string
    type: object
  item.AddRecyclingCenterRequest:
    properties:
      accepted_materials:
        items:
          type: string
        type: array
      address:
        type: string
      contact_number:
        type: string
      name:
        type: string
      working_hours:
        type: string
    type: object
...
  item.GetRatingsRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
      user_id:
        type: string
    type: object
  item.GetRatingsResponse:
    properties:
      average_rating:
        type: number
      limit:
        type: integer
      page:
        type: integer
      ratings:
        items:
          $ref: '#/definitions/item.Rating'
        type: array
      total_ratings:
        type: integer
    type: object
  item.GetStatisticsRequest:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    type: object
  item.GetStatisticsResponse:
    properties:
      top_categories:
        items:
          $ref: '#/definitions/item.TopCategory'
        type: array
      top_recycling_centers:
        items:
          $ref: '#/definitions/item.TopRecyclingCenter'
        type: array
      total_eco_points_earned:
        type: integer
      total_recycled_items:
        type: integer
      total_swaps:
        type: integer
    type: object
...
  item.ListRecyclingCentersResponse:
    properties:
      centers:
        items:
          $ref: '#/definitions/item.RecyclingCenterResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  item.ListSwapRequestsRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
      status:
        type: string
    type: object
  item.ListSwapRequestsResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      swaps:
        items:
          $ref: '#/definitions/item.SwapResponse'
        type: array
      total:
        type: integer
    type: object
...
  item.Rating:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      rater_id:
        type: string
      rating:
        type: number
      swap_id:
        type: string
    type: object
  item.RecyclingCenterResponse:
    properties:
      accepted_materials:
        items:
          type: string
        type: array
      address:
        type: string
      contact_number:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      working_hours:
        type: string
    type: object
  item.RecyclingItem:
    properties:
      item_id:
        type: string
      material:
        type: string
      weight:
        type: number
    type: object
  item.RecyclingSubmissionResponse:
    properties:
      center_id:
        type: string
      created_at:
        type: string
      eco_points_earned:
        type: integer
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/item.RecyclingItem'
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  item.RejectSwapRequestRequest:
    properties:
      reason:
        type: string
      swap_id:
        type: string
    type: object
...
  item.SearchRecyclingCentersRequest:
    properties:
      limit:
        type: integer
      material:
        type: string
      page:
        type: integer
    type: object
  item.SendSwapRequestRequest:
    properties:
      message:
        type: string
      offered_item_id:
        type: string
      requested_item_id:
        type: string
      user_id:
        type: string
    type: object
  item.SubmitItemsForRecyclingRequest:
    properties:
      center_id:
        type: string
      items:
        items:
          $ref: '#/definitions/item.RecyclingItem'
        type: array
      user_id:
        type: string
    type: object
  item.SwapResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      offered_item_id:
        type: string
      reason:
        type: string
      requested_item_id:
        type: string
      requester_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  item.TopCategory:
    properties:
      id:
        type: string
      name:
        type: string
      swap_count:
        type: integer
    type: object
  item.TopRecyclingCenter:
    properties:
      id:
        type: string
      name:
        type: string
      submissions_count:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway of User Item System
  title: User Item System
  version: "1.0"
paths:
...
  /item-system/ratings/GetAll:
    post:
      description: Retrieves all ratings info from ratings table in PostgreSQL
      parameters:
      - description: rating data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.GetRatingsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.GetRatingsResponse'
        "500":
          description: Server error while getting ratings
          schema:
            type: string
      summary: Gets all ratings
      tags:
      - rating
  /item-system/ratings/add:
    post:
      description: Inserts new rating info into ratings table in PostgreSQL
      parameters:
      - description: New rating data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.AddRatingRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.Rating'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while adding rating
          schema:
            type: string
      summary: Adds a new rating
      tags:
      - rating
  /item-system/recycling:
    post:
      description: Inserts recycling submission info into the database
      parameters:
      - description: New recycling submission data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.SubmitItemsForRecyclingRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.RecyclingSubmissionResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while submitting items for recycling
          schema:
            type: string
      summary: Submits items for recycling
      tags:
      - recycling
  /item-system/recycling-centers:
    post:
      description: Inserts new recycling center info into the database
      parameters:
      - description: New recycling center data
        in: body
        name: new_data
        required: true
        schema:
          $ref: '#/definitions/item.AddRecyclingCenterRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.RecyclingCenterResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error while adding recycling center
          schema:
            type: string
      summary: Adds a new recycling center
      tags:
      - recycling_center
  /item-system/recycling-centers/search:
    post:
      description: Retrieves recycling centers based on search criteria
      parameters:
      - description: Search criteria
        in: body
        name: search_criteria
        required: true
        schema:
          $ref: '#/definitions/item.SearchRecyclingCentersRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ListRecyclingCentersResponse'
        "400":
          description: Invalid search criteria
          schema:
            type: string
        "500":
          description: Server error while searching recycling centers
          schema:
            type: string
      summary: Searches for recycling centers
      tags:
      - recycling_center
  /item-system/statistics:
    post:
      description: Retrieves various statistics from the service
      parameters:
      - description: Statistics filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/item.GetStatisticsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.GetStatisticsResponse'
        "400":
          description: Invalid filter
          schema:
            type: string
        "500":
          description: Server error while getting statistics
          schema:
            type: string
      summary: Gets statistics
      tags:
      - statistics
  /item-system/swaps:
    post:
      description: Sends a swap request to the service
      parameters:
      - description: Swap request info
        in: body
        name: swap
        required: true
        schema:
          $ref: '#/definitions/item.SendSwapRequestRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.SwapResponse'
        "400":
          description: Invalid request data
          schema:
            type: string
        "500":
          description: Server error while sending swap request
          schema:
            type: string
      summary: Send swap request
      tags:
      - swap
  /item-system/swaps/accept:
    put:
      description: Accepts a swap request
      parameters:
      - description: Swap request info
        in: body
        name: swap
        required: true
        schema:
          $ref: '#/definitions/item.AcceptSwapRequestRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.SwapResponse'
        "400":
          description: Invalid request data
          schema:
            type: string
        "500":
          description: Server error while accepting swap request
          schema:
            type: string
      summary: Accept swap request
      tags:
      - swap
  /item-system/swaps/list:
    post:
      description: Lists all swap requests
      parameters:
      - description: Swap request filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/item.ListSwapRequestsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.ListSwapRequestsResponse'
        "400":
          description: Invalid filter
          schema:
            type: string
        "500":
          description: Server error while listing swap requests
          schema:
            type: string
      summary: List swap requests
      tags:
      - swap
  /item-system/swaps/reject:
    put:
      description: Rejects a swap request
      parameters:
      - description: Swap request info
        in: body
        name: swap
        required: true
        schema:
          $ref: '#/definitions/item.RejectSwapRequestRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/item.SwapResponse'
        "400":
          description: Invalid request data
          schema:
            type: string
        "500":
          description: Server error while rejecting swap request
          schema:
            type: string
      summary: Reject swap request
      tags:
      - swap
...
swagger: "2.0"